+++ Changelog.md
@@ -5,6 +5,7 @@
 - Updated NEventStore 9.0.0.
 - Added support for net6.0.
 - Updated MongoDB driver to 2.12.4.
+- Configuration: allow to configure MongoClientSettings to edit driver specific client connection settings [#60](https://github.com/NEventStore/NEventStore.Persistence.MongoDB/issues/60).
 
 ## 8.0.0